requirements.txt
luigi==2.1.1
boto==2.40.0
boto3==1.3.1
pyyaml==3.11